+++ setup.cfg
@@ -37,7 +37,6 @@ python_requires = >=3.6
 where = .
 include =
     tmaven
-	tmaven.bin
     tmaven.*
 
 [options.package_data]
@@ -45,4 +44,4 @@ tmaven = *.png, *.svg
 
 [options.entry_points]
 console_scripts =
-	tmaven=tmaven.bin.tmaven_gui:main
+	tmaven=tmaven.__main__:main
